job_board/pyproject.toml
[project]
name = "job-board"
version = "0.1.0"
description = ""
authors = [
    {name = "Drew",email = "ayobamioduola13@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djoser (>=2.3.1,<3.0.0)"

]

[tool.poetry]
packages = [{include = "job_board", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
